Do not run most cookbooks when node is in readying state
The readying state is the state that the node has when booting. Since
the cookbooks ensure that everything is configured properly (including
for reboots), we do not need convergence to enforce the config in that
case.

We do this for dns-server (not dns-client yet!), logging, nfs_client,
ntp, provisioner, suse-manager-client and updater.

The main risk is for provisioner, but the configuration should be on
disk already.

We do not do this yet for the roles that run in all states (crowbar,
deployer-client, dns-client), nor for ipmi-discover (where we likely
want some ipmi check on boot), nor for network (this needs some more
attention, as there's code to kill NICs not managed by crowbar on boot).