utils: Make role recipes ignore role if node is applying another role
REWRITE ME HACK FOR HA COMPUTE ONLY

"When we apply a proposal, we actually apply several batches of roles.
For each batch, we look at what roles are applied and only apply these.

The only exception is deployer-client, as it's required both for
heartbeat and setting up some utilities for recipes.